requirements.txt
cryptography==38.0.1
Flask==2.2.2
PyJWT==2.5.0
requests==2.28.1
